move text styling version to index
